operator tasks

+++ src/ex_10_Switch/Lab094_JDK13_Multicase.java
@@ -2,8 +2,8 @@ package ex_10_Switch;
 
 public class Lab094_JDK13_Multicase {
     public static void main(String[] args) {
-        int itemCode = 002;
-        switch(itemCode){
+        long itemCode = 002;
+        switch((int) itemCode){
             case 001,002,005:
                 System.out.println("All of the them are Electronic Gadget");
                 System.out.println("All of the them are Electronic Gadget");
